explaining why descrepancy happens
The answer given regarding "DNS resolutions" for this question doesn't say why it happens, nor that it is in fact is quite common and expected on today's internet, for DNS to return different IPs according to the geographical location of the querying user, which may be puzzling to many, judging by the level of the answers given indicating a completely non-technical audience. 
The proposed text should also help clarify the explanation of "geographical distribution of content".